Fix import statement.
The Zeek-specific part of the analyzers imported the DHCP module with a
non-existing path. This was incorrect ever since this analyzer was moved
out of Spicy where the path existed into zeek/spicy-analyzers, and then
also when it got moved to its own plugin. It appears this only worked
because Spicy did not handle import statements with scopes correctly
which (for at least what is happening here) is not the case anymore once
zeek/spicy#1187 is fixed.

This patch uses an unscoped import.