Make test `analyzer.dpd` compatible with >=zeek-4.2.
This patch adjusts the baseline of the test `analyzer.dpd` so that it
passes with both zeek-4.0 as well as >zeek-4.2. Between the two version
the ordering of reported lines has changed (previously would have logged
`#close` before the log record) and the `host` column now also includes
the port.

We drop comment lines and the `host` column.